+++ src/sll/init.c
@@ -71,9 +71,9 @@ static void _write_stack_frame(sll_file_descriptor_t fd,sll_instruction_index_t 
 
 __SLL_NO_RETURN void _force_exit(const sll_char_t* a,const sll_char_t* b,const sll_char_t* c){
 	sll_file_descriptor_t fd=sll_platform_get_default_stream_descriptor(SLL_PLATFORM_STREAM_ERROR);
-	sll_platform_file_write(fd,a,sll_string_length_unaligned(a),NULL);
-	sll_platform_file_write(fd,b,sll_string_length_unaligned(b),NULL);
-	sll_platform_file_write(fd,c,sll_string_length_unaligned(c),NULL);
+	sll_platform_file_write(fd,a,sll_string_length(a),NULL);
+	sll_platform_file_write(fd,b,sll_string_length(b),NULL);
+	sll_platform_file_write(fd,c,sll_string_length(c),NULL);
 	if (sll_current_runtime_data&&sll_current_runtime_data->a_dt){
 		const sll_call_stack_t* c_st=sll_thread_get_call_stack(_scheduler_current_thread_index);
 		if (c_st){