+++ CHANGELOG.md
@@ -35,6 +35,7 @@
 - Reduced the number of calls to `sll_platform_lock_acquire` and `sll_platform_lock_release`
 - Reduced the number of thread-local variables
 - Renamed `SLL_ERROR_FLAG_LIBC` and `error_codes$LIBC_ERROR_CODES` to `SLL_ERROR_FLAG_SYSTEM` and `error_codes$SYSTEM_ERROR_CODES`
+- Renamed `sll_string_length_unaligned` to `sll_string_length`
 - Renamed `XXX_fail` function to `XXX_raw`
 - Renamed argument parsing modifier symbolizing variadic functions from `+` to `!`
 - Rewritten the CLI executable loader code
@@ -46,6 +47,7 @@
 
 - Thread number limit macro: `SLL_SCHEDULER_MAX_THREADS`
 - Unused define: `SLL_ERROR_FLAG_WINAPI`
+- Unused function: `sll_string_length`
 - Unused map: `error$WINDOWS_ERROR_CODES`
 - Unused type: `sll_decl_copy_node_data_t`